3_grpc_server_stream/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
requests==2.31.0 